+++ lib/src/models/pass_json/location/location.dart
@@ -1,26 +1,43 @@
+import 'package:equatable/equatable.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'location.g.dart';
 
+/// Locations where the pass is relevant.
 @JsonSerializable()
-class Location {
+class Location extends Equatable {
   /// Optional. Altitude, in meters, of the location.
-  double altitude;
+  final double altitude;
 
   /// Required. Latitude, in degrees, of the location.
-  double latitude;
+  final double latitude;
 
   /// Required. Longitude, in degrees, of the location.
-  double longitude;
+  final double longitude;
 
   /// Optional. Text displayed on the lock screen when the pass is currently relevant.
   /// For example, a description of the nearby location such as “Store nearby on 1st and Main.”
-  String relevantText;
+  final String relevantText;
 
-  Location({this.altitude, this.latitude, this.longitude, this.relevantText});
+  Location({
+    this.altitude,
+    this.latitude,
+    this.longitude,
+    this.relevantText,
+  });
 
+  /// Convert from json
   factory Location.fromJson(Map<String, dynamic> json) =>
       _$LocationFromJson(json);
 
+  /// Convert to json
   Map<String, dynamic> toJson() => _$LocationToJson(this);
+
+  @override
+  List<Object> get props => [
+        this.altitude,
+        this.latitude,
+        this.longitude,
+        this.relevantText,
+      ];
 }