+++ lib/src/models/pass_json/pass_json.g.dart
@@ -59,9 +59,9 @@ PassJson _$PassJsonFromJson(Map<String, dynamic> json) {
         ?.map((e) =>
             e == null ? null : Location.fromJson(e as Map<String, dynamic>))
         ?.toList(),
-  )
-    ..maxDistance = json['maxDistance'] as int
-    ..relevantDate = json['relevantDate'] as String;
+    maxDistance: json['maxDistance'] as int,
+    relevantDate: json['relevantDate'] as String,
+  );
 }
 
 Map<String, dynamic> _$PassJsonToJson(PassJson instance) => <String, dynamic>{