+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## 1.0.1
+## 1.0.2 - 28.03.2020
+ - fix bug with maxDistance and relevantDate in pass json
+ - fix bug with auxiliaryFields, primaryFields, transitType
+ - add equatable for comparisons pass.json data
+ - add documentation for Fields
+
+## 1.0.1 - 13.03.2020
  - fix code format
  - add altitude to Location Dictionary Keys
  - add maxDistance to Relevance Keys
@@ -7,5 +13,5 @@
  - add documentation to Pass Structure Dictionary model
  - add documentation for Pass Json
 
-## 1.0.0
+## 1.0.0 - 13.03.2020
  - initial release